Get organization teams

get_teams returns a (filtered) data frame of teams in the organization with columns for their names (name) and their unique ids (id).

get_teams(org, filter = NULL, exclude = FALSE)

Arguments

org
character, name of the GitHub organization.

filter
character, regex pattern for matching (or excluding) repos.

exclude
logical, should entries matching the regex be excluded or included.

See also

Examples

# NOT RUN {
get_team_repos("ghclass",c("team01","team02"))
# }