ghclass: A package for github based classroom and assignment management

ghclass: A package for github based classroom and assignment management

github repo functions

One -

Two -

local repo functions

clone_repo - clone repos locally

pull_repo - update local repos via pull

github organization functions

One -

Two -

wercker functions

One -

Two -

authentication functions

GitHub

get_github_token - get github token

set_github_token - set github token

test_github_token - test github token

Wercker

get_wercker_token - get wercker token

set_wercker_token - set wercker token

test_wercker_token - test wercker token