GitHub API related tools

github_whoami returns the login of the authenticated user.

github_get_api_limit / github_set_api_limit - get and set the limit of results to return from the GitHub API.

github_get_token returns the user's GitHub personal access token (PAT).

github_set_token defines the user's GitHub PAT by setting the GITHUB_PAT enivronmental variable. This value can then

github_reset_token removes the value stored in the GITHUB_PAT environmental variable.

github_test_token checks if a PAT is valid by attempting to authenticate with the GitHub API

github_get_api_limit()

github_set_api_limit(limit = 1000L)

github_get_token()

github_set_token(token)

github_reset_token()

github_test_token(token = github_get_token())

github_whoami(quiet = FALSE)

Arguments

limit
The maximum number of records to return from an API request.

token
Character. Either the literal token path, or the path of a file containing the token.

quiet
Logical. Should status messages be shown.

Details

This package looks for the personal access token (PAT) in the following places (in order):

Value of GITHUB_PAT environmental variable.

Value of GITHUB_TOKEN environmental variable.

Contents of ~/.github/token file.

See also

Examples

# whoami
github_whoami()

#> [1] "rundel"

# Set the API request limit
github_get_api_limit()

#> [1] 1000

github_set_api_limit(500)
github_get_api_limit()

#> [1] 500

# Set and get the PAT
github_set_token("~/.github/token")
github_set_token("0123456789ABCDEF")
pat = github_get_token()
pat

#> [1] "0123456789ABCDEF"

# Test the PAT
github_reset_token()
try(github_test_token())

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> Error : 

try(github_test_token("bad_token"))

#> x Your GitHub PAT failed to authenticate.

#> API message: Bad credentials

#> API docs: https://developer.github.com/v3

#> Missing page:

#> 
#> └─GitHub API error (401): 401 Unauthorized
#>   ├─
#>   └─