GitHub Issue related tools

issue_create creates a new issue.

issue_close closes an existing issue.

issue_edit edits the properties of an existing issue.

Usage

issue_close(repo, number)

issue_create(repo, title, body, labels = character(), assignees = character())

issue_edit(
  repo,
  number,
  title = NULL,
  body = NULL,
  state = NULL,
  milestone = NULL,
  labels = list(),
  assignees = list()
)

Arguments

repo

Character. Address of one or more repositories in owner/name format.

number

Integer. GitHub issue number.

title

Character. Title of the issue.

body

Character. Content of the issue.

labels

Character. Vector of the labels to associate with this issue

assignees

Character. Vector of logins for users assigned to the issue.

state

Character. State of the issue. Either "open" or "closed".

milestone

Character. The number of the milestone to associate this issue with. Only users with push access can set the milestone for issues. The milestone is silently dropped otherwise.

See also

Examples

# \dontrun{
repo_create("ghclass-test","test_issue")
#> ✔ Created repo "ghclass-test/test_issue".

issue_create(
  "ghclass-test/test_issue",
  title = "Issue 1",
  body = "This is an issue"
 )
#> ✖ Failed to create issue "Issue 1" for repo "ghclass-test/test_issue".
#> └─GitHub API error (404): 
#>   ├─ API message: Not Found
#>   ├─ API docs: https://docs.github.com/rest/reference/issues#create-an-issue
#>   └─ Missing page: https://api.github.com/repos/ghclass-test/test_issue/issues

issue_create(
  "ghclass-test/test_issue",
  title = "Issue 2", body = "This is also issue",
  label = "Important"
)
#> ✖ Failed to create issue "Issue 2" for repo "ghclass-test/test_issue".
#> └─GitHub API error (404): 
#>   ├─ API message: Not Found
#>   ├─ API docs: https://docs.github.com/rest/reference/issues#create-an-issue
#>   └─ Missing page: https://api.github.com/repos/ghclass-test/test_issue/issues

issue_create(
  "ghclass-test/test_issue",
  title = "Issue 3", body = "This is also issue",
  label = c("Important", "Super Important"),
  assignees = "rundel"
)
#> ✖ Failed to create issue "Issue 3" for repo "ghclass-test/test_issue".
#> └─GitHub API error (404): 
#>   ├─ API message: Not Found
#>   ├─ API docs: https://docs.github.com/rest/reference/issues#create-an-issue
#>   └─ Missing page: https://api.github.com/repos/ghclass-test/test_issue/issues

issue_close("ghclass-test/test_issue", 1)
#> ✖ Failed to close issue "#1" for repo "ghclass-test/test_issue".
#> └─GitHub API error (404): 
#>   ├─ API message: Not Found
#>   ├─ API docs: https://docs.github.com/rest/reference/issues#update-an-issue
#>   └─ Missing page: https://api.github.com/repos/ghclass-test/test_issue/issues/1

issue_edit(
  "ghclass-test/test_issue", 2,
  title = "New issue 2 title!",
  body = "Replacement body text"
)
#> ✖ Failed to edit issue #2 for repo "ghclass-test/test_issue".
#> └─GitHub API error (404): 
#>   ├─ API message: Not Found
#>   ├─ API docs: https://docs.github.com/rest/reference/issues#update-an-issue
#>   └─ Missing page: https://api.github.com/repos/ghclass-test/test_issue/issues/2

ghclass::repo_issues("ghclass-test/test_issue")
#> ✖ Failed to retrieve issues for repo "ghclass-test/test_issue".
#> └─GitHub API error (404): 
#>   ├─ API message: Not Found
#>   ├─ API docs: https://docs.github.com/rest/reference/issues#list-repository-iss
#>   └─ Missing page: https://api.github.com/repos/ghclass-test/test_issue/issues?s
#> # A tibble: 0 × 0

repo_delete("ghclass-test/test_issue", prompt=FALSE)
#> ✔ Deleted repo "ghclass-test/test_issue".
# }