GitHub Organization related tools

org_sitrep - Provides a situation report on a GitHub organization.

org_set_repo_permission - Change the default permission level for org repositories.

org_admins - returns a vector of repository administrators. In the case of a non-organization owner (e.g. a user account) returns the owner's login.

org_exists - returns TRUE if the organization(s) exist on GitHub and FALSE otherwise.

org_invite - invites user(s) to a GitHub organization.

org_remove - remove user(s) from an organization (and all teams within that organization).

org_members - returns a (filtered) vector of organization members.

org_pending - returns a (filtered) vector of pending organization members.

org_repos - returns a (filtered) vector of organization repositories.

org_teams - returns a (filtered) vector of organization teams.

org_team_details - returns a data frame of all organization teams containing identification and permission details.

Experimental

org_repo_stats - returns a tibble of repositories belonging to a GitHub organization along with some basic statistics about those repositories.

org_admins(org)

org_exists(org)

org_invite(org, user)

org_members(org, filter = NULL, exclude = FALSE, include_admins = TRUE)

org_pending(org, filter = NULL, exclude = FALSE)

org_remove(org, user, prompt = TRUE)

org_repo_stats(
  org,
  filter = "",
  filter_type = "in:name",
  inc_commits = TRUE,
  inc_issues = TRUE,
  inc_prs = TRUE
)

org_repos(org, filter = NULL, exclude = FALSE, full_repo = TRUE)

org_sitrep(org)

org_team_details(org)

org_teams(org, filter = NULL, exclude = FALSE, team_type = c("name", "slug"))

org_set_repo_permission(org, permission = c("none", "read", "write", "admin"))

Arguments

org
Character. Name of the GitHub organization(s).

user
Character. GitHub username(s).

filter
Character. Regular expression pattern for matching (or excluding) results

exclude
Logical. Should entries matching the regular expression be excluded or included.

include_admins
Logical. Should admin users be included in the results.

prompt
Logical. Prompt before removing member from organization.

filter_type
Character. One or more GitHub search in qualifiers. See documentation for more details.

inc_commits
Logical. Include commit statistics

inc_issues
Logical. Include issue statistics

inc_prs
Logical. Include pull request statistics

full_repo
Logical. Should the full repository address be returned (e.g. owner/repo instead of just repo)?

team_type
Character. Either "slug" if the team names are slugs or "name" if full team names are provided.

permission
Default permission level members have for organization repositories:

read - can pull, but not push to or administer this repository.

write - can pull and push, but not administer this repository.

admin - can pull, push, and administer this repository.

none - no permissions granted by default.

See also

Examples

if (FALSE) {
# Org Details
org_admins("ghclass-test")
org_admins("rundel") # User, not an organization

org_exists(c("rundel","ghclass-test"))

# Org Membership - Invite, Status, and Remove
students = c("ghclass-anya", "ghclass-bruno", "ghclass-celine",
         "ghclass-diego", "ghclass-elijah","ghclass-francis")
org_invite(students)

org_members("ghclass-test")
org_pending("ghclass-test")

org_remove("ghclass-test", students)
org_pending("ghclass-test")

# Org repos and teams
org_repos("ghclass-test")
org_repos("ghclass-test", "hw1-")
org_teams("ghclass-test")

org_team_details("ghclass-test")
}

if (FALSE) {
org_repo_stats("ghclass-test")
}