Create author rating form

peer_form_create_rating creates a short feedback form for authors to rate the feedback they got from reviewers. The rating categories are based on Reily, K. and P. Ludford Finnerty, and L. Terveen (2009): Two Peers Are Better Than One: Aggregating Peer Reviews for Computing Assignments is Surprisingly Accurate. In Proceedings of the ACM 2009 International Conference on Supporting Group Work. GROUP’09, May 10–13, 2009, Sanibel Island, Florida, USA.

peer_form_create_rating(
  category = c("helpfulness", "accuracy", "fairness"),
  title = "Author rating form",
  fname = "feedback_blank_rating",
  output = "github_document",
  write_rmd = TRUE,
  overwrite = FALSE,
  allow_comment = FALSE
)

Arguments

category
Character. Categories to be included in the feedback form, defaults to c("helpfulness", "accuracy", "fairness").

title
Character. Title of form, defaults to "Author rating form".

fname
Character. File name of RMarkdown document to be written to memory, defaults to feedback_blank_rating.

output
Character. Output parameter for .Rmd file, defaults to github_document.

write_rmd
Logical. Whether the feedback form should be saved to a .Rmd file in the current working directory, defaults to TRUE.

overwrite
Logical. Should existing file or files with same name be overwritten, defaults to FALSE.

allow_comment
Logical. Should optional comment field be included? Defaults to FALSE.

See also

Examples

if (FALSE) {
peer_form_create_rating(c("accuracy", "fairness"))
}