Create peer review roster

peer_roster_create creates data frame of random assignments of author files to reviewers. By default, the output is saved to a .csv file in the current working directory that incorporates the current date and random seed as part of the file name.

peer_roster_create(n_rev, user, seed = NULL, write_csv = TRUE, dir = NULL)

Arguments

n_rev
Numeric. Number of reviews per user. Must be larger than zero and smaller than the number of users.

user
Character. A vector of GitHub user names.

seed
Numeric. Random seed for assignment, if not set a random seed will be chosen.

write_csv
Logical. Whether the roster data frame should be written to a .csv file in the current working directory, defaults to TRUE.

dir
Character. Directory where the peer review roster will be written if write_csv = TRUE.

See also

Examples

if (FALSE) {
peer_roster_create(
  3,
  c("anya-ghclass", "bruno-ghclass",
    "celine-ghclass", "diego-ghclass"),
  dir = "hw2/"
)
}