Retrieve information about GitHub Actions workflows and their runs.

action_workflows() - retrieve details on repo workflows.

action_runs() - retrieve details on repo workflow runs.

action_status() - DEPRECATED - retrieve details on most recent workflow runs.

action_runtime() - retrieves runtime durations for workflow runs.

action_artifacts() - retrieve details on available workflow artifacts.

action_artifact_download() - downloads artifact(s) into a local directory.

action_artifact_delete() - deletes artifact(s).

Usage

action_artifacts(repo, keep_expired = FALSE, which = c("latest", "all"))

action_artifact_delete(repo, ids)

action_artifact_download(
  repo,
  dir,
  ids = action_artifacts(repo),
  keep_zip = FALSE,
  file_pat = "",
  overwrite = FALSE
)

action_runs(
  repo,
  branch = NULL,
  event = NULL,
  status = NULL,
  created = NULL,
  limit = 1
)

action_status(
  repo,
  branch = NULL,
  event = NULL,
  status = NULL,
  created = NULL,
  limit = 1
)

action_runtime(
  repo,
  branch = NULL,
  event = NULL,
  status = NULL,
  created = NULL,
  limit = 1
)

action_workflows(repo, full = FALSE)

Arguments

repo

Character. Address of repository in owner/name format.

keep_expired

Logical. Should expired artifacts be returned.

which

Character. Either "latest" to return only the most recent of each artifact or "all" to return all artifacts.

ids

Integer or data frame. Artifact ids to be downloaded or deleted. If a data frame is passed then the id column will be used.

dir

Character. Path to the directory where artifacts will be saved.

keep_zip

Logical. Should the artifact zips be saved (TRUE) or their contents (FALSE).

file_pat

Character. If extracting zip with multiple files, regexp pattern to match filename.

overwrite

Logical. Should existing files be overwritten.

branch

Character. Filter runs associated with a particular branch.

event

Character. Filter runs for triggered by a specific event. See here for possible event names.

status

Character. Filter runs for a particular status or conclusion (e.g. completed or success).

created

Character. Filter runs for a given creation date. See here for date query syntax.

limit

Numeric. Maximum number of workflow runs to return. Default 1. Note results are chronologically ordered, so limit = 1 will return the most recent action run for a repository.

full

Logical. Should all workflow columns be returned. Default FALSE.

Value

action_workflows(), action_runs(), action_runtime(), and action_artifacts all return tibbles containing information on requested repos' available workflows, recent workflow runs, workflow runs runtimes, and generated artifacts respectively.

action_artifact_download() returns a character vector containing the paths of all downloaded fules

action_artifact_delete() returns an invisible data frame containing repository names and ids of the deleted artifacts.

Examples

if (FALSE) { # \dontrun{
action_workflows("rundel/ghclass")

action_runs("rundel/ghclass")

action_runtime(c("rundel/ghclass", "rundel/parsermd"))

action_artifacts(c("rundel/ghclass", "rundel/parsermd"))
} # }