Tools for limiting gh's GitHub api requests.

github_get_api_limit() - returns the current limit on results returned by gh.

github_set_api_limit() - sets a limit on results returned by gh.

Usage

github_get_api_limit()

github_set_api_limit(limit = 10000L)

Arguments

limit

The maximum number of records to return from an API request.

Value

github_get_api_limit() returns a single integer value.

github_set_api_limit() invisibily returns the value of the limit argument.

Details

This value is stored in the "ghclass.api.limit" option globally.

Examples

github_get_api_limit()
#> [1] 10000

github_set_api_limit(500)

github_get_api_limit()
#> [1] 500