GitHub Organization related tools

org_admins - returns a vector of repository administrators. In the case of a non-organization owner (e.g. a user account) returns the owner's login.

org_exists - returns TRUE if the organization(s) exist on GitHub and FALSE otherwise.

org_invite - invites user(s) to a GitHub organization.

org_remove - remove user(s) from an organization (and all teams within that organization).

org_members - returns a (filtered) vector of organization members.

org_pending - returns a (filtered) vector of pending organization members.

org_repos - returns a (filtered) vector of organization repositories.

org_teams - returns a (filtered) vector of organization teams.

org_team_details - returns a data frame of all organization teams containing identification and permission details.

Experimental

org_repo_stats - returns a tibble of repositories belonging to a GitHub organization along with some basic statistics about those repositories.

org_admins(org)

org_exists(org)

org_invite(org, user)

org_members(org, filter = NULL, exclude = FALSE, include_admins = TRUE)

org_pending(org, filter = NULL, exclude = FALSE)

org_remove(org, user, prompt = TRUE)

org_repo_stats(
  org,
  filter = "",
  filter_type = "in:name",
  inc_commits = TRUE,
  inc_issues = TRUE,
  inc_prs = TRUE
)

org_repos(org, filter = NULL, exclude = FALSE, full_repo = TRUE)

org_team_details(org)

org_teams(org, filter = NULL, exclude = FALSE, team_type = c("slug", "name"))

Arguments

org
Character. Name of the GitHub organization(s).

user
Character. GitHub username(s).

filter
Character. Regular expression pattern for matching (or excluding) results

exclude
Logical. Should entries matching the regular expression be excluded or included.

include_admins
Logical. Should admin users be included in the results.

prompt
Logical. Prompt before removing member from organization.

filter_type
Character. One or more GitHub search in qualifiers. See documentation for more details.

inc_commits
Logical. Include commit statistics

inc_issues
Logical. Include issue statistics

inc_prs
Logical. Include pull request statistics

full_repo
Logical. Should the full repository address be returned (e.g. owner/repo instead of just repo)?

team_type
Character. Either "slug" if the team names are slugs or "name" if full team names are provided.

See also

Examples

# \dontrun{
# Org Details
org_admins("ghclass-test")

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> Failed to retrieve admins for org 'ghclass-test'.

#> Error: 

org_admins("rundel") # User, not an organization

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> Failed to retrieve admins for org 'rundel'.

#> Error: 

org_exists(c("rundel","ghclass-test"))

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> [1] FALSE FALSE

# Org Membership - Invite, Status, and Remove
students = c("ghclass-anya", "ghclass-bruno", "ghclass-celine",
         "ghclass-diego", "ghclass-elijah","ghclass-francis")
org_invite(students)

#> Argument 'org' must be of length 1.

#> Error: 

org_members("ghclass-test")

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> character(0)

org_pending("ghclass-test")

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> x Failed to retrieve pending members for org 'ghclass-test'

#> API message:

#> API docs:

#> Missing page:

#> 
#> └─

#> character(0)

org_remove("ghclass-test", students)

#> Error in eval(expr, envir = list(`?` = function(...) stop()), enclos = envir): object 'x' not found

org_pending("ghclass-test")

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> x Failed to retrieve pending members for org 'ghclass-test'

#> API message:

#> API docs:

#> Missing page:

#> 
#> └─

#> character(0)

# Org repos and teams
org_repos("ghclass-test")

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> Organization 'ghclass-test' does not exist on GitHub.

#> x Failed to retrieve repos for org 'ghclass-test'.

#> API message:

#> API docs:

#> Missing page:

#> 
#> └─

org_repos("ghclass-test", "hw1-")

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> Organization 'ghclass-test' does not exist on GitHub.

#> x Failed to retrieve repos for org 'ghclass-test'.

#> API message:

#> API docs:

#> Missing page:

#> 
#> └─

org_teams("ghclass-test")

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> x Failed to retrieve teams for org 'ghclass-test'.

#> API message:

#> API docs:

#> Missing page:

#> 
#> └─

#> character(0)

org_team_details("ghclass-test")

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> x Failed to retrieve teams for org 'ghclass-test'.

#> API message:

#> API docs:

#> Missing page:

#> 
#> └─

#> # A tibble: 0 x 5
#> # … with 5 variables: slug <chr>, id <int>, name <chr>, privacy <chr>,
#> #   permission <chr>

# }

# \dontrun{
org_repo_stats("ghclass-test")

#> This function ('org_repo_stats') is currently considered experimental. Its
#> interface, implementation, and other features may change significantly in
#> future versions of the package. Use with caution.

#> Warning: 

#> Unable to locate a github token, please use `github_set_token()` or define the
#> GITHUB_PAT environmental variable in your .Renviron file.

#> Error: 

# }