Create a team or individual assignment

This is a higher level function that combines the following steps:

Create repos

Create teams and invite students if necessary

Add teams or individuals to the repositories

Mirror a template repository to assignment repositories

org_create_assignment(
  org,
  repo,
  user,
  team = NULL,
  source_repo = NULL,
  private = TRUE
)

Arguments

org
Character. Name of the GitHub organization.

repo
Character. Name of the repo(s) for the assignment.

user
Character. GitHub username(s).

team
Character. Team names, if not provided an individual assignment will be created.

source_repo
Character. Address of the repository to use as a template for all created repos.

private
Logical. Should the created repositories be private.