Obtain details on an organization's repos and teams

org_exists() - returns TRUE if the organization(s) exist on GitHub and FALSE otherwise.

org_teams() - returns a (filtered) vector of organization teams.

org_team_details() - returns a data frame of all organization teams containing identification and permission details.

org_repos() - returns a (filtered) vector of organization repositories.

org_repo_search() - search for repositories within an organization (preferred for large organizations).

org_repo_stats() - returns a tibble of repositories belonging to a GitHub organization along with some basic statistics about those repositories.

Usage

org_exists(org)

org_repo_search(org, name, extra = "", full_repo = TRUE)

org_repo_stats(
  org,
  branch = NULL,
  filter = "",
  filter_type = "in:name",
  inc_commits = TRUE,
  inc_issues = TRUE,
  inc_prs = TRUE
)

org_repos(
  org,
  filter = NULL,
  exclude = FALSE,
  full_repo = TRUE,
  sort = c("full_name", "created", "updated", "pushed"),
  direction = c("asc", "desc"),
  type = c("all", "public", "private", "forks", "sources", "member", "internal")
)

org_team_details(org)

org_teams(org, filter = NULL, exclude = FALSE, team_type = c("name", "slug"))

Arguments

org

Character. Name of the GitHub organization(s).

name

Character. Full or partial repo name to search for within the org

extra

Character. Any additional search qualifiers, see Searching for repositories for details.

full_repo

Logical. Should the full repository address be returned (e.g. owner/repo instead of just repo).

branch

Character. The branch to use for counting commits, if NULL then each repo's default branch is used.

filter

Character. Regular expression pattern for matching (or excluding) results

filter_type

Character. One or more GitHub search in qualifiers. See documentation for more details.

inc_commits

Logical. Include commit statistics (branch, commits, last_update)

inc_issues

Logical. Include issue statistics (open_issues, closed_issues)

inc_prs

Logical. Include pull request statistics (open_prs, merged_prs, closed_prs)

exclude

Logical. Should entries matching the regular expression be excluded or included.

sort

Character. Sorting criteria to use, can be one of "created", "updated", "pushed", or "full_name".

direction

Character. Sorting order to use.

type

Character. Specifies the type of repositories you want, can be one of "all", "public", "private", "forks", "sources", "member", or "internal".

team_type

Character. Either "slug" if the team names are slugs or "name" if full team names are provided.

Value

org_exists() returns a logical vector.

org_teams(), org_repos, and org_repo_search() return a character vector of team or repo names.

org_team_details() and org_repo_stats() return tibbles.

Examples

if (FALSE) { # \dontrun{
# Org repos and teams
org_repos("ghclass-test")

org_repos("ghclass-test", filter = "hw1-")

org_teams("ghclass-test")

org_team_details("ghclass-test")
} # }