Organization permissions

org_sitrep - Provides a situation report on a GitHub organization.

org_set_repo_permission - Change the default permission level for org repositories.

org_sitrep(org)

org_set_repo_permission(org, permission = c("none", "read", "write", "admin"))

Arguments

org
Character. Name of the GitHub organization(s).

permission
Default permission level members have for organization repositories:

read - can pull, but not push to or administer this repository.

write - can pull and push, but not administer this repository.

admin - can pull, push, and administer this repository.

none - no permissions granted by default.

Examples

# \dontrun{
org_sitrep("ghclass-test")

#> 

#> ── ghclass-test sitrep: ─────────────────────────────────────────────────────────────────────────────────

#> Public repos: 0

#> ●

#> Private repos: 1

#> ●

#> Members: 0

#> ●

#> Default repository permission: 'none'

#> ●

#> Members can create public repos: FALSE

#> ●

#> Members can create private repos: FALSE

#> ●

org_set_repo_permission("ghclass-test", "read")

#> ✓ Set org 'ghclass-test''s repo permissions to 'read'.

org_sitrep("ghclass-test")

#> 

#> ── ghclass-test sitrep: ─────────────────────────────────────────────────────────────────────────────────

#> Public repos: 0

#> ●

#> Private repos: 1

#> ●

#> Members: 0

#> ●

#> Default repository permission: read

#> ●

#>   ℹ Warning - members can currently view all repos in this org.

#> Members can create public repos: FALSE

#> ●

#> Members can create private repos: FALSE

#> ●

# Cleanup
org_set_repo_permission("ghclass-test", "none")

#> ✓ Set org 'ghclass-test''s repo permissions to 'none'.

# }