GitHub Repository tools - repository details

repo_clone_url() - Returns the url, for cloning, a GitHub repo (either ssh or https)

repo_branches() - Returns a (filtered) vector of branch names.

repo_commits() - Returns a tibble of commits to a GitHub repository.

repo_issues() - Returns a tibble of issues for a GitHub repository.

repo_n_commits() - Returns a tibble of the number of commits in a GitHub repository (and branch).

repo_prs() - Returns a tibble of pull requests for a GitHub repository.

Usage

repo_branches(repo)

repo_clone_url(repo, type = c("https", "ssh"))

repo_commits(
  repo,
  branch = NULL,
  sha = branch,
  path = NULL,
  author = NULL,
  since = NULL,
  until = NULL,
  quiet = FALSE
)

repo_issues(
  repo,
  state = c("open", "closed", "all"),
  assignee = NULL,
  creator = NULL,
  mentioned = NULL,
  labels = NULL,
  sort = c("created", "updated", "comments"),
  direction = c("desc", "asc"),
  since = NULL
)

repo_n_commits(repo, quiet = FALSE)

repo_prs(repo, state = c("open", "closed", "all"))

Arguments

repo

Character. Address of repository in owner/name format.

type

Character. Clone url type, either "https" or "ssh".

branch

Character. Branch to list commits from.

sha

Character. SHA to start listing commits from.

path

Character. Only commits containing this file path will be returned.

author

Character. GitHub login or email address by which to filter commit author.

since

Character. Only issues updated at or after this time are returned.

until

Character. Only commits before this date will be returned, expects YYYY-MM-DDTHH:MM:SSZ format.

quiet

Logical. Should an error message be printed if a repo does not exist.

state

Character. Pull request state.

assignee

Character. Return issues assigned to a particular username. Pass in "none" for issues with no assigned user, and "*" for issues assigned to any user.

creator

Character. Return issues created the by the given username.

mentioned

Character. Return issues that mentioned the given username.

labels

Character. Return issues labeled with one or more of of the given label names.

sort

Character. What to sort results by. Can be either "created", "updated", or "comments".

direction

Character. The direction of the sort. Can be either "asc" or "desc".

Value

repo_clone_url() and repo_branches() both return a character vector. repo_commits(), repo_issues(), repo_n_commits(), and repo_prs() all return a tibble.

Examples

# \dontrun{
repo_clone_url("rundel/ghclass")
#> [1] "https://github.com/rundel/ghclass.git"

repo_branches("rundel/ghclass")
#> [1] "gert_update"   "git_add_all"   "github_sitrep" "master"       
#> [5] "peer_review"   "testing"      

repo_commits("rundel/ghclass")
#> # A tibble: 500 × 6
#>    repo           login  name         email            date                msg  
#>    <chr>          <chr>  <chr>        <chr>            <dttm>              <chr>
#>  1 rundel/ghclass rundel Colin Rundel rundel@gmail.com 2022-01-04 21:14:04 Smal…
#>  2 rundel/ghclass rundel Colin Rundel rundel@gmail.com 2022-01-04 21:09:02 cran…
#>  3 rundel/ghclass rundel Colin Rundel rundel@gmail.com 2022-01-04 21:08:53 Doc …
#>  4 rundel/ghclass rundel Colin Rundel rundel@gmail.com 2022-01-04 21:07:26 User…
#>  5 rundel/ghclass rundel Colin Rundel rundel@gmail.com 2022-01-04 21:06:05 Team…
#>  6 rundel/ghclass rundel Colin Rundel rundel@gmail.com 2022-01-04 17:58:30 retu…
#>  7 rundel/ghclass rundel Colin Rundel rundel@gmail.com 2022-01-04 17:58:18 repo…
#>  8 rundel/ghclass rundel Colin Rundel rundel@gmail.com 2022-01-04 17:23:44 pr f…
#>  9 rundel/ghclass rundel Colin Rundel rundel@gmail.com 2022-01-04 17:22:47 org …
#> 10 rundel/ghclass rundel Colin Rundel rundel@gmail.com 2022-01-04 17:14:19 Loca…
#> # … with 490 more rows

repo_issues("rundel/ghclass")
#> ✔ Retrieved issues for repo "rundel/ghclass".
#> # A tibble: 33 × 10
#>    repo   number title    state comments created             updated            
#>    <chr>   <int> <chr>    <chr>    <int> <dttm>              <dttm>             
#>  1 runde…    126 Zoom/Mo… open         1 2021-01-23 19:48:29 2021-01-23 19:51:33
#>  2 runde…    123 repo_co… open         0 2020-11-26 14:24:28 2020-11-26 14:24:28
#>  3 runde…    121 peer_in… open         3 2020-11-23 03:38:39 2020-11-30 05:13:14
#>  4 runde…    119 org_rep… open         1 2020-10-29 17:31:00 2020-10-29 20:24:42
#>  5 runde…    117 Example… open         0 2020-10-21 21:54:54 2020-10-21 21:54:54
#>  6 runde…    116 Duplica… open         0 2020-10-21 12:20:52 2020-10-21 12:20:52
#>  7 runde…    115 Does th… open         4 2020-10-18 17:10:16 2020-10-23 09:04:07
#>  8 runde…    106 New ste… open         0 2020-04-08 00:10:38 2020-04-08 00:10:38
#>  9 runde…    105 Peer-re… open         0 2020-03-30 10:06:30 2020-03-30 10:06:30
#> 10 runde…    103 org_rep… open         1 2020-02-15 17:07:29 2020-02-17 11:31:14
#> # … with 23 more rows, and 3 more variables: closed <dttm>, created_by <chr>,
#> #   assignees <list>

repo_n_commits("rundel/ghclass", branch = "master")
#> Error in repo_n_commits("rundel/ghclass", branch = "master"): unused argument (branch = "master")

repo_prs("rundel/ghclass")
#> # A tibble: 1 × 7
#>   repo              pr        id title                   state base_ref head_ref
#>   <chr>          <int>     <int> <chr>                   <chr> <chr>    <chr>   
#> 1 rundel/ghclass   106 400561584 New steps to apply for… open  master   patch-1 
# }