GitHub Repository tools - repository details

repo_clone_url - Returns the url, for cloning, a GitHub repo (either ssh or https)

repo_branches - Returns a (filtered) vector of branch names.

repo_commits - Returns a tibble of commits to a GitHub repository.

repo_issues - Returns a tibble of issues for a GitHub repository.

repo_n_commits - Returns a tibble of the number of commits in a GitHub repository (and branch).

repo_prs - Returns a tibble of pull requests for a GitHub repository.

repo_branches(repo)

repo_clone_url(repo, type = c("https", "ssh"))

repo_commits(
  repo,
  branch = NULL,
  sha = branch,
  path = NULL,
  author = NULL,
  since = NULL,
  until = NULL,
  quiet = FALSE
)

repo_issues(
  repo,
  state = c("open", "closed", "all"),
  assignee = NULL,
  creator = NULL,
  mentioned = NULL,
  labels = NULL,
  sort = c("created", "updated", "comments"),
  direction = c("desc", "asc"),
  since = NULL
)

repo_n_commits(repo, quiet = FALSE)

repo_prs(repo, state = c("open", "closed", "all"))

Arguments

repo
Character. Address of repository in owner/name format.

type
Character. Clone url type, either "https" or "ssh".

branch
Character. Branch to list commits from.

sha
Character. SHA to start listing commits from.

path
Character. Only commits containing this file path will be returned.

author
Character. GitHub login or email address by which to filter commit author.

since
Character. Only issues updated at or after this time are returned.

until
Character. Only commits before this date will be returned, expects YYYY-MM-DDTHH:MM:SSZ format.

quiet
Logical. Should an error message be printed if a repo does not exist.

state
Character. Pull request state.

assignee
Character. Return issues assigned to a particular username. Pass in "none" for issues with no assigned user, and "*" for issues assigned to any user.

creator
Character. Return issues created the by the given username.

mentioned
Character. Return issues that mentioned the given username.

labels
Character. Return issues labeled with one or more of of the given label names.

sort
Character. What to sort results by. Can be either "created", "updated", or "comments".

direction
Character. The direction of the sort. Can be either "asc" or "desc".

Examples

# \dontrun{
repo_clone_url("rundel/ghclass")

#> [1] "https://github.com/rundel/ghclass.git"

repo_branches("rundel/ghclass")

#> [1] "gert_update" "master"      "peer_review"

repo_commits("rundel/ghclass")

#> # A tibble: 500 x 6
#>    repo     login  name   email       date                msg                   
#>    <chr>    <chr>  <chr>  <chr>       <dttm>              <chr>                 
#>  1 rundel/… rundel Colin… rundel@gma… 2020-10-08 13:56:06 "Merge branch 'master…
#>  2 rundel/… rundel Colin… rundel@gma… 2020-10-08 13:55:26 "Possible fix for #11…
#>  3 rundel/… rundel Colin… rundel@gma… 2020-10-08 13:54:54 "Remove dependency on…
#>  4 rundel/… rundel Colin… rundel@gma… 2020-10-07 15:51:27 "Handle the case of n…
#>  5 rundel/… NA     runner runner@Mac… 2020-10-07 12:26:50 "Re-build README.Rmd" 
#>  6 rundel/… rundel Colin… rundel@gma… 2020-10-07 12:22:08 "Update vignette for …
#>  7 rundel/… rundel Colin… rundel@gma… 2020-10-01 09:47:32 "pkgdown"             
#>  8 rundel/… rundel Colin… rundel@gma… 2020-10-01 09:46:00 "Add explicit permiss…
#>  9 rundel/… rundel Colin… rundel@gma… 2020-09-23 11:10:54 "Merge branch 'master…
#> 10 rundel/… rundel Colin… rundel@gma… 2020-09-23 11:09:52 "Hide dep warning"    
#> # … with 490 more rows

repo_issues("rundel/ghclass")

#> ✓ Retrieved issues for repo 'rundel/ghclass'.

#> # A tibble: 28 x 10
#>    repo  number title state comments created             updated            
#>    <chr>  <int> <chr> <chr>    <int> <dttm>              <dttm>             
#>  1 rund…    114 org_… open         9 2020-10-08 12:57:45 2020-10-08 14:46:27
#>  2 rund…    113 Erro… open         0 2020-10-07 15:37:35 2020-10-07 15:37:35
#>  3 rund…    106 New … open         0 2020-04-08 00:10:38 2020-04-08 00:10:38
#>  4 rund…    105 Peer… open         0 2020-03-30 10:06:30 2020-03-30 10:06:30
#>  5 rund…    103 org_… open         1 2020-02-15 17:07:29 2020-02-17 11:31:14
#>  6 rund…     97 Test… open         1 2020-01-31 00:57:02 2020-02-11 13:13:38
#>  7 rund…     95 org_… open         1 2020-01-28 19:35:58 2020-01-28 19:41:12
#>  8 rund…     89 `pee… open         0 2019-11-06 23:53:04 2019-11-06 23:53:04
#>  9 rund…     88 `org… open         0 2019-11-06 20:25:22 2019-11-06 20:25:22
#> 10 rund…     87 `pee… open         0 2019-11-03 15:51:44 2019-11-06 23:49:11
#> # … with 18 more rows, and 3 more variables: closed <dttm>, created_by <chr>,
#> #   assignees <list>

repo_n_commits("rundel/ghclass", branch = "master")

#> Error in repo_n_commits("rundel/ghclass", branch = "master"): unused argument (branch = "master")

repo_prs("rundel/ghclass")

#> # A tibble: 1 x 7
#>   repo          pr       id title                        state base_ref head_ref
#>   <chr>      <int>    <int> <chr>                        <chr> <chr>    <chr>   
#> 1 rundel/gh…   106   4.01e8 New steps to apply for GH o… open  master   patch-1 

# }